+++ packages/terra-clinical-label-value-view/src/terra-dev-site/test/clinical-label-value-view/TextValueLabelValueView.test.jsx
@@ -8,7 +8,6 @@ const cx = classNames.bind(styles);
 
 const examples = () => (
   <div>
-    <h3>LabelValueView returned as a term and a definition (isChildOfDescriptionList=&apos;true&apos;):</h3>
     <style>
       {`
         #dl-wrapper > dd {
@@ -16,7 +15,9 @@ const examples = () => (
         }
       `}
     </style>
-    <dl id="dl-wrapper">
+    <h3>LabelValueView returned as a term and a definition (isChildOfDescriptionList=&apos;true&apos;):</h3>
+    <h4 id="vitalsTopic">Vitals Topic Text</h4>
+    <dl id="dl-wrapper" aria-labelledby="vitalsTopic">
       <LabelValueView label="Collected By" textValue="Noah Brown" isChildOfDescriptionList />
       <LabelValueView label="Date" textValue="12/12/12 00:00AM" isChildOfDescriptionList />
       <LabelValueView label="Blood Pressure" textValue="110/60 mmHg" isChildOfDescriptionList />
@@ -24,12 +25,11 @@ const examples = () => (
       <LabelValueView label="Heart Rate" textValue="80 BPM" isChildOfDescriptionList />
     </dl>
     <h3>LabelValueView returned as a description list (isChildOfDescriptionList=&apos;false&apos;):</h3>
-    <LabelValueView label="Collected By" textValue="Noah Brown" isChildOfDescriptionList={false} />
-    <LabelValueView label="Date" textValue="12/12/12 00:00AM" isChildOfDescriptionList={false} />
+    <h4 id="collectedByTopic">Collected By Topic Text</h4>
+    <LabelValueView label="Collected By" textValue="Noah Brown" isChildOfDescriptionList={false} ariaLabelledBy="collectedByTopic" />
     <h3>LabelValueView returned as a description list (isChildOfDescriptionList is not given - will take the default value false):</h3>
-    <LabelValueView label="Blood Pressure" textValue="110/60 mmHg" />
-    <LabelValueView label="Temperature" textValue="97 F" />
-    <LabelValueView label="Heart Rate" textValue="80 BPM" />
+    <h4 id="temperatureTopic">Temperature Topic Text</h4>
+    <LabelValueView label="Temperature" textValue="97 F" ariaLabelledBy="temperatureTopic" />
     <h3>Applied max-width of 200px to show word wrapping:</h3>
     <div className={cx('labelvalueview-wrapper')}>
       <LabelValueView id="LongLabelValueView" label="supercalifragilisticexpialidocious" textValue="supercalifragilisticexpialidocious" />